Initial version of demo graph
This commit includes not only the graph itself, but also the scripts
that should be used to start dlg, submit the graph, and finally shut dlg
down. The graph execution itself is still failing due to a few factors
(NGAS drops not supported by the translation process and by the LG UI,
filenames not fully compatible with how modules currently work,
not-fully-working modules, etc), but at least it is correctly carried
out.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>