NGAS plugin to react on file archivals
This plug-in should be configured as an archiving event handler plug-in
on the subscribing NGAS server (i.e., the one receiving data from the
site). In order to be able to load it, the server should also be
configured to load python modules from the directory where this plug-in
is (see the NGAS documentation for that).

The plug-in will run as the NGAS user, so care must be taken and
the plug-in modified accordingly (e.g., sudo into other user, ensure
that PATH is properly set, etc).

This is a first approach to solving FAST-47.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>